requirements.txt
Django == 1.8.17
djangorestframework
django-allauth
pep8
django-cors-headers